feat: support add new route to typescript config when add a block (#2435)
* fix: block dependencies target error

* feat: support add new route to typescript config when add a block

* feat: support add a block to a ts page

* feat: support insert new block component in typescript file

* remove console.log

* test: add more test case for block

+++ packages/umi-build-dev/src/plugins/commands/block/appendBlockToContainer.js
@@ -3,7 +3,7 @@ import insertComponent from './insertComponent';
 
 const debug = require('debug')('umi-build-dev:appendBlockToContainer');
 
-export default ({ entryPath, blockFolderName }) => {
+export default ({ entryPath, blockFolderName, dryRun }) => {
   debug('start to update the entry file for block(s) under the path...');
 
   const oldEntry = readFileSync(entryPath, 'utf-8');
@@ -12,5 +12,7 @@ export default ({ entryPath, blockFolderName }) => {
     relativePath: `./${blockFolderName}`,
   });
 
-  writeFileSync(entryPath, newEntry);
+  if (!dryRun) {
+    writeFileSync(entryPath, newEntry);
+  }
 };
